feat: refactor project structure to use 'src/' directory and modern TypeScript practices
- Moved all source code into 'src/' directory for better organization.
- Updated TypeScript configuration to reflect new directory structure.
- Added CLI entry point in 'src/cli/main.ts' for knowledge graph operations.
- Modularized code by separating types, graph management, and server functionality.
- Implemented new server functionality with tool registration and request handling.
- Added tests for graph manager and server functionalities.
- Updated package.json to include necessary dependencies and scripts for testing.
