+++ makefile
@@ -1,10 +1,13 @@
-.PHONY: build run npx-build clean
+.PHONY: build run npx-build clean test
 
 build:
 	npm run build
 
 run:
 	npm start
+
+test:
+	npm test
 
 npx-build: build
 	npm pack
