+++ src/graph/index.ts
@@ -1,0 +1,2 @@
+// Export all graph functionality
+export * from './manager.js';
