+++ src/index.ts
@@ -1,0 +1,6 @@
+// Export all components from a single entry point
+export * from './types/index.js';
+export * from './graph/index.js';
+export * from './server/index.js';
+
+// Note: CLI isn't exported as it's an entrypoint
