+++ src/server/index.ts
@@ -1,0 +1,2 @@
+// Export all server functionality
+export * from './mcp-server.js';
