+++ src/types/index.ts
@@ -1,0 +1,2 @@
+// Export all type definitions
+export * from './knowledge-graph.js';
