+++ tsconfig.json
@@ -2,9 +2,9 @@
     "extends": "./tsconfig.base.json",
     "compilerOptions": {
       "outDir": "./dist",
-      "rootDir": "."
+      "rootDir": "src"
     },
     "include": [
-      "./**/*.ts"
+      "src/**/*.ts"
     ]
 }
